requirement.txt
selenium==4.33.0
pytest==8.3.5
Appium-Python-Client==5.1.1
allure-pytest==2.14.3
PyYAML==6.0.2